[Props] Add another property "TypeEx"
dleyna-server used to "overload" the Type property with new possible
values in a way that made it incompatible with other MediaServer2Spec
users. Add a new property TypeEx that includes this "overloaded"
implementation so that Type can be left as specified.

dleyna-server clients that used to depend on this old Type-
functionality should start using TypeEx.

Fixes #26.

Signed-off-by: Jussi Kukkonen <jussi.kukkonen@intel.com>