[Bug Fix] Use object id instead of the parent id for detecting root object
Use object id instead of the parent id to detect if we are at the root
object or not.
This workaround the case where root have an empty string as parent id
value but avoid to return an invalid object path.
According to the specification root parent id must have a "-1" value.

Fix issue: https://github.com/01org/dleyna-server/issues/32

Signed-off-by: Christophe Guiraud <christophe.guiraud@intel.com>