[Device] Add sleeping device memory cache
 when the last context of a sleeping device is notified as
unavailable, instead of having the device and its context deleted,
the device is kept alive, the lost server notification is not
broadcasted, the device is moved from the "device_udn_map" to a
"sleeping device udn" map and the context is stored
in a new device structure context field "sleeping_device_context".
All pending operations related to the device will also be cancelled.

The GetServer() request returns now the devices available in the
"sleeping device udn" map in addition to those in the "device udn map".

when a get_prop request is done on a device with a
sleeping_device_context available this one is used to perform the
request.

all Dbus requests other than "get_prop" and "wake" invoked on a device
with a sleeping_device_context available fail with an error
"DLEYNA_ERROR_OPERATION_FAILED".

when a new device proxy is notified as available, if it belongs to the
"sleeping device udn" map, then it is removed from the
"sleeping device udn" map and detroyed, the lost server notification
is then broadcasted. The new proxy device is then constructed as usual.

Signed-off-by: Christophe Guiraud <christophe.guiraud@intel.com>