[Device] Move the network interface info storage
The network interface info which is used for the device wake-on
is now stored in the dls_device_t structure instead of the
dls_device_context_t structure.
This makes more sense since we only store the one network interface
info element (the best matching) whatever the device context count
of a device.

Signed-off-by: Christophe Guiraud <christophe.guiraud@intel.com>