[Path] Make object paths stable across restarts
Fix issue #119: <https://github.com/01org/dleyna-server/issues/119>

Root paths for servers were builded using an incremental counter,
based of the discovery order made by GSSDP:

Results were of this form:

/com/intel/dLeynaServer/server/0
/com/intel/dLeynaServer/server/1

Object paths were builded on these root paths:

/com/intel/dLeynaServer/server/0/xxxxxxxxxx

There is no persistence between restarts of dleyna because we can't
assign the same index to the same server.

To fix this issue, we replace the volatile part of the root path, id the index,
by a non volatile and unique part, id the server udn.

The server udn is a string of this form: "uuid:<uuid string>"
<uuid string> is defined by RFC 4122(<http://www.ietf.org/rfc/rfc4122.txt>
and it's composed only with hex digit and char '-'

To make the path compatible with dbus and 'reversible':
1 - remove the prefix 'uuid:'
2 - convert '-' char to '_' char

Result is of the form below:

/com/intel/dLeynaServer/server/99951d71_23d5_5525_94eb_a4aa78a98211

This path is stable accross dleyna restart and identify uniquely a
server. This can be sued to save object path and to use them after
multiple restart of the server and dLeyna.

Signed-off-by: Ludovic Ferrandis <ludovic.ferrandis@intel.com>